+++ changelog.md
@@ -12,6 +12,7 @@ Release 0.13.2
 - Fixed spamming chat messages for reeling/exhausted.
 - Fixed the 'user created equipment' flag on equipment creation.
 - Added system setting to turn on/off chat messages for reeling/exhausted.
+- Added warning for bad sound file
 
 Release 0.13.1 1/12/2022